@@ -7,13 +7,10 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"runtime"
-
-	"github.com/kr/binarydist"
+	"time"
 )
 
 var version, genDir string
@@ -21,11 +18,13 @@ var version, genDir string
 type current struct {
 	Version string
 	Sha256  []byte
+	Channel string
+	Date    time.Time
 }
 
 func generateSha256(path string) []byte {
 	h := sha256.New()
-	b, err := ioutil.ReadFile(path)
+	b, err := os.ReadFile(path)
 	if err != nil {
 		fmt.Println(err)
 	}
@@ -35,38 +34,14 @@ func generateSha256(path string) []byte {
 	//return base64.URLEncoding.EncodeToString(sum)
 }
 
-type gzReader struct {
-	z, r io.ReadCloser
-}
-
-func (g *gzReader) Read(p []byte) (int, error) {
-	return g.z.Read(p)
-}
-
-func (g *gzReader) Close() error {
-	g.z.Close()
-	return g.r.Close()
-}
-
-func newGzReader(r io.ReadCloser) io.ReadCloser {
-	var err error
-	g := new(gzReader)
-	g.r = r
-	g.z, err = gzip.NewReader(r)
-	if err != nil {
-		panic(err)
-	}
-	return g
-}
-
-func createUpdate(path string, platform string) {
-	c := current{Version: version, Sha256: generateSha256(path)}
+func createUpdate(path string, platform string, channel string) {
+	c := current{Version: version, Sha256: generateSha256(path), Channel: channel, Date: time.Now()}
 
 	b, err := json.MarshalIndent(c, "", "    ")
 	if err != nil {
 		fmt.Println("error:", err)
 	}
-	err = ioutil.WriteFile(filepath.Join(genDir, platform+".json"), b, 0755)
+	err = os.WriteFile(filepath.Join(genDir, platform+".json"), b, 0755)
 	if err != nil {
 		panic(err)
 	}
@@ -75,53 +50,18 @@ func createUpdate(path string, platform string) {
 
 	var buf bytes.Buffer
 	w := gzip.NewWriter(&buf)
-	f, err := ioutil.ReadFile(path)
+	f, err := os.ReadFile(path)
 	if err != nil {
 		panic(err)
 	}
 	w.Write(f)
 	w.Close() // You must close this first to flush the bytes to the buffer.
-	err = ioutil.WriteFile(filepath.Join(genDir, version, platform+".gz"), buf.Bytes(), 0755)
+	err = os.WriteFile(filepath.Join(genDir, version, platform+".gz"), buf.Bytes(), 0755)
 
-	files, err := ioutil.ReadDir(genDir)
 	if err != nil {
-		fmt.Println(err)
+		panic(err)
 	}
 
-	for _, file := range files {
-		if file.IsDir() == false {
-			continue
-		}
-		if file.Name() == version {
-			continue
-		}
-
-		os.Mkdir(filepath.Join(genDir, file.Name(), version), 0755)
-
-		fName := filepath.Join(genDir, file.Name(), platform+".gz")
-		old, err := os.Open(fName)
-		if err != nil {
-			// Don't have an old release for this os/arch, continue on
-			continue
-		}
-
-		fName = filepath.Join(genDir, version, platform+".gz")
-		newF, err := os.Open(fName)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Can't open %s: error: %s\n", fName, err)
-			os.Exit(1)
-		}
-
-		ar := newGzReader(old)
-		defer ar.Close()
-		br := newGzReader(newF)
-		defer br.Close()
-		patch := new(bytes.Buffer)
-		if err := binarydist.Diff(ar, br, patch); err != nil {
-			panic(err)
-		}
-		ioutil.WriteFile(filepath.Join(genDir, file.Name(), version, platform), patch.Bytes(), 0755)
-	}
 }
 
 func printUsage() {
@@ -159,6 +99,7 @@ func main() {
 	platform := *platformFlag
 	appPath := flag.Arg(0)
 	version = flag.Arg(1)
+	channel := flag.Arg(2)
 	genDir = *outputDirFlag
 
 	createBuildDir()
@@ -170,14 +111,14 @@ func main() {
 	}
 
 	if fi.IsDir() {
-		files, err := ioutil.ReadDir(appPath)
+		files, err := os.ReadDir(appPath)
 		if err == nil {
 			for _, file := range files {
-				createUpdate(filepath.Join(appPath, file.Name()), file.Name())
+				createUpdate(filepath.Join(appPath, file.Name()), file.Name(), channel)
 			}
 			os.Exit(0)
 		}
 	}
 
-	createUpdate(appPath, platform)
+	createUpdate(appPath, platform, channel)
 }
