@@ -3,7 +3,7 @@ package main
 import (
 	"log"
 
-	"github.com/sanbornm/go-selfupdate/selfupdate"
+	"gitlab.com/wingbits/backend/go-libraries/go-selfupdate/selfupdate"
 )
 
 // The purpose of this app is to provide a simple example that just prints
